fix channel tests and move implementations
